pyproject.toml
[project]
name = "django-content-studio"
version = "1.0.0-beta.5"
description = "Modern & flexible Django admin"
authors = [
    {name = "Leon van der Grient",email = "leon@devtastic.io"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[tool.poetry]
packages = [{include = "content_studio"}]
include = [
    { path = "content_studio/static/**/*", format = "sdist" },
    { path = "content_studio/static/**/*", format = "wheel" }]

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
Django= "^5.2.9"
djangorestframework= "^3.16.1"
djangorestframework-simplejwt= "^5.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
